+++ src/app/søknader/omsorgspenger-utbetaling/schema.ts
@@ -1,8 +1,15 @@
 import { get } from 'lodash';
 
 import { IPeriode } from 'app/models/types';
-import yup, { barn, passertDato, passertKlokkeslettPaaMottattDato, periode, utenlandsperiode } from 'app/validation/yup';
+import yup, {
+    barn,
+    passertDato,
+    passertKlokkeslettPaaMottattDato,
+    periode,
+    utenlandsperiode,
+} from 'app/validation/yup';
 import { erYngreEnn4år } from 'app/utils';
+import { isDateBefore } from 'app/utils/date/dateUtils';
 
 import nb from '../../i18n/nb.json';
 import { IOMPUTSoknad } from './types/OMPUTSoknad';
@@ -136,7 +143,14 @@ const frilanser = () =>
             .string()
             .when('jobberFortsattSomFrilans', {
                 is: false,
-                then: (schema) => schema.required('Sluttdato er et påkrevd felt.'),
+                then: (schema) =>
+                    schema
+                        .required('Sluttdato er et påkrevd felt.')
+                        .test(
+                            'sluttdato-etter-startdato',
+                            'Sluttdato kan ikke være før startdato.',
+                            (sluttdato, context) => !isDateBefore(sluttdato, context.parent.startdato),
+                        ),
             })
             .label('Sluttdato'),
         jobberFortsattSomFrilans: yup.boolean(),